Detail Changes
Added button link to Instagram.
Play around with color but decided on the origional black and white because of the simplicity of it. Changed the link "Photos" to "Gallery"
shifted things around more so that the site would look better in mobile browser. However having extreme difficulty formating it in the case of the use of laptop.